// src/app/modules/administrator/manage-master/model/manage-master.model.ts
export class MasterTypeModel {
  description!: string;
  displayName!: string;
  name!: string;
  sequence!: number;
  sysMasterId!: number;
  sysMasterTableId!: number;
  api?: string;
  primaryKey?: string;
  deleteString?: string;
  compositeKey?: string;
  showName?: string;
  getString?: string;
}